address ming's comment
